http: move Req from module root to its own file
